feat: IssueComment 생성 로직 및 깃허브 api 로직 추가
- tagId 및 Count 필드 변수 타입을 Integer로 변경 = 8바이트 -> 4바이트
- Queue를 이용해 오류 발생 시 재시도 로직을 추가함(OOM 가능성 존재하므로 나중에 최적화할 예정)
- CommentCommand에 어노테이션 변경- Getter와 SuperBuilder 조합
- 깃허브 api를 통해 pull과 comment를 요청하는 로직 추가

+++ api/src/main/java/com/opensourcereader/api/client/response/GithubIssueResponse.java
@@ -6,12 +6,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record GithubIssueResponse(
     Long id,
-    @JsonProperty("number") Long tagId,
+    @JsonProperty("number") Integer tagId,
     String title,
     @JsonProperty("user") GithubUserResponse user,
     String body,
     String state,
-    @JsonProperty("comments") Long commentCount,
+    @JsonProperty("comments") Integer commentCount,
     @JsonProperty("created_at") Instant createdAt,
     @JsonProperty("updated_at") Instant updatedAt,
     @JsonProperty("pull_request") GithubPullRequestInfo pullRequest) {
